strutil/shlex,osutil/udev/netlink: minimally import go-check
This allows us to use -gocheck options for the whole codebase, for example to
run all tests in the codebase with the name matching "kernel" one can now do:

go test ./... -gocheck.f kernel

and that will only fail if the tests fail, whereas previously it would fail
because in these packages gocheck was not imported and as such -gocheck.f was
not recognized as an option and the help text for `go test` was printed off
instead of running any tests.

Eventually we may want to adapt the tests in these packages to use go-check too
so that these can also be filtered/run successfully with go test and gocheck
options.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>