chore: update minimum Go version to 1.18 (#24)
Copying / contributing packages of interest to x-go is difficult since
the baseline version of Go in the other projects has moved on from
x-go's current version of 1.14. Let's reasonably raise the baseline so
we can add some functionality.

The current Go version for the primary projects of interest to x-go is
as follows:

- Snapd (not using x-go yet): 1.18
- Chisel (not using x-go yet): 1.23.8
- Pebble (using x-go): 1.24.6

The current Go version for other projects which import x-go is as follows:

- Concierge (using x-go): 1.24.4

Let's update x-go to match the minimum Go version of snapd (1.18).