feat: introduce testutil package based on pebble (#25)
Copy the internals/testutil package from Pebble at commit
b58ea3bf3755c50792154c2380a874d379601cc0.

Many of the internal projects have grown a `testutil` package, which, as
its name implies, supports common testing setup and utilities. While
this package is itself in the spirit of "common packages" across our
projects we would like to consolidate into x-go, it's also the case that
many of the other packages we may be interested in consolidating here
are also reliant on the testing functionality that `testutil` provides.
To facilitate copying further packages of interest, we start first with
`testutil`.

Additionally, we decide to start with the version of `testutil` found in
Pebble as it is:

1. The most common subset of this package across the projects which use
a version of it.
2. The most recently updated version of this subset

---

This branch is based on #24 due to the usage of `any` and generics in
the copied code.

The commands used generate the changes here (copy code + modify imports)
are in the description of each respective commit.