feat: introduce osutil package based on pebble (#26)
Copy the internals/osutil package from Pebble at commit
b58ea3bf3755c50792154c2380a874d379601cc0, except for the osutil/squashfs
package.

The rationale for this is very much like that given in #25. Many of our
projects now have an `osutil` package, and the version of `osutil` in
Pebble seems to be the most modern version of the same subset of code
from snapd. Let's start by copying this code over and then we can update
and make additions to the code as we like from here.

We intentionally drop the osutil/squashfs package as the use-case for
that code is likely too narrow. We should revisit this in the future if
it becomes more clear this isn't the case.